+++ tests/integration-test-samples/validation/1/non_boolean_value_float.php
@@ -22,5 +22,5 @@ $expectedErrors = [
 $ares = new Ares($schema);
 
 $this->assertSame(empty($expectedErrors), $ares->validate($data));
-$this->assertEquals($expectedErrors, $ares->getValidationErrors());
+$this->assertEquals($expectedErrors, $ares->getValidationErrors()->getArrayCopy());