+++ tests/integration-test-samples/validation/122/custom_decimal_type.php
@@ -31,5 +31,5 @@ $expectedErrors = [
 $ares = new Ares($schema);
 
 $this->assertSame(empty($expectedErrors), $ares->validate($data));
-$this->assertEquals($expectedErrors, $ares->getValidationErrors());
+$this->assertEquals($expectedErrors, $ares->getValidationErrors()->getArrayCopy());