+++ tests/integration-test-samples/validation/122/custom_validation_rule.php
@@ -45,4 +45,4 @@ $expectedErrors = [
 $ares = new Ares($schema);
 
 $this->assertSame(empty($expectedErrors), $ares->validate($data));
-$this->assertEquals($expectedErrors, $ares->getValidationErrors());
+$this->assertEquals($expectedErrors, $ares->getValidationErrors()->getArrayCopy());