+++ CHANGELOG.md
@@ -4,11 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+### Added
+### Fixed
+
+## [4.0.0] 2020-10-18
 ### Added
 - regex validation rule pattern check and error handling
+- ```Ares\Ares::getValidationErrors()``` now returns an error collection object capable of generating json:api style and nested array error collections
 ### Fixed
 - fixes clean code issues
+- updates outdated ```README.md``` contents
 
 ## [3.0.0] 2020-08-16
 ### Added